% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_Support_Tools.R
\name{DE_Support_Tools}
\alias{DE_Support_Tools}
\alias{level_matcher}
\alias{UpdateLevelsbyList}
\alias{valid_factors}
\title{Tools to support DE analysis}
\usage{
level_matcher(raw_levels = NULL, ordered_levels = NULL)

UpdateLevelsbyList(metadata = NULL, named_levels = NULL)

valid_factors(metadata = NULL)
}
\arguments{
\item{raw_levels}{string of factor levels in data to be sorted.}

\item{ordered_levels}{string of factor levels in the desired order}

\item{metadata}{A data.frame containing factors}

\item{named_levels}{a named list containing the
levels to set as reference via
\code{\link[forcats]{fct_relevel}}}
}
\value{
data.frame with factors
}
\description{
Tools to support DE analysis
}
\details{
\code{\link[=level_matcher]{level_matcher()}} returns overlapping strings between
raw_levels and ordered_levels, with the ordering of the later.
No partial string matching. This is useful when applying a generic
factor level order to different metadata that do not share
all the same levels.

The names from named_levels map to the
columns in metadata to update. List slots
contain strings passed to the ... argument of
\code{\link[forcats]{fct_relevel}}

This is all done with \code{\link[purrr]{imap_dfc}}

drops single level factors with
 metadata %>%
 select_if(~n_distinct(na.omit(.x))>1)
}
