% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{DEphe}
\alias{DEphe}
\title{Heatmap generator}
\usage{
DEphe(
  Hotgenes = NULL,
  ExpressionSlots = NULL,
  SampleIDs = NULL,
  hotList = NULL,
  contrasts = NULL,
  padj_cut = 0.1,
  .log2FoldChange = 0,
  Topn = Inf,
  annotations = NULL,
  mode = "all",
  border_color = "white",
  scale = "row",
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  arrangeby = NULL,
  label_by = "Feature",
  ...
)
}
\arguments{
\item{Hotgenes}{Hotgenes object.}

\item{ExpressionSlots}{name of normalized data to use.
If NULL (default), the first name returned by ExpressionSlots_
will be used.
See HotgenesObj.}

\item{SampleIDs}{vector of sample ids to select}

\item{hotList}{vector of Features to select. This will overide
all cut offs.}

\item{contrasts}{vector for DE data.frames in Output_DE slot to return.
See ?contrasts_. If NULL (default), all contrasts used.}

\item{padj_cut}{numeric value for padj limit (0.1 is Default).}

\item{.log2FoldChange}{absolute log2FoldChange for filtering.
Default is 0.}

\item{Topn}{numeric value for the maximum number of DEGs to show.
Arranged by padj. Default is Inf for all available DEGs.}

\item{annotations, }{string indicating the coldata_ids to select.
If NULL (default), none are returned (unlike all with coldata_).}

\item{mode}{string indicating the type of coldata to call.
Choices include: "all", "quanti", "quali". If "all", the
default, all columns will be returned. If "quanti", only
numeric coldata will be returned. If "quali", only
qualitative coldata will be returned.}

\item{border_color}{color of cell borders on heatmap, use NA if no border should be 
drawn.}

\item{scale}{character indicating if the values should be centered and scaled in 
either the row direction or the column direction, or none. Corresponding values are 
\code{"row"}, \code{"column"} and \code{"none"}}

\item{cluster_rows}{boolean values determining if rows should be clustered or \code{hclust} object,}

\item{cluster_cols}{boolean values determining if columns should be clustered or \code{hclust} object.}

\item{arrangeby}{String mapping to a column name returned by
coldata_. Samples in the heatmap will be sorted by this variable.}

\item{label_by}{String matching a column name returned by
Mapper_ function, default is "Feature".}

\item{...}{Additional parameters for
\code{\link[pheatmap]{pheatmap}}.}
}
\description{
Heatmap generator
}
\details{
If Feature maps to multiple names, only one will be
used.
}
\examples{
if(interactive()) {
  # load packages
  library(Hotgenes)
  
  dds_Hotgenes_dir <- system.file(
    "extdata",
    paste0("dds_Hotgenes", ".RDS"),
    package = "Hotgenes",
    mustWork = TRUE
  )
  # from DESeq2
  htgs <- readRDS(dds_Hotgenes_dir)
  
  # DEphe -------------------------------------------------------------------
  ss <- SampleIDs_(htgs)[1:8]
  ss
  
  feat <- Features_(htgs)
   
  
  htgs \%>\%
    DEphe(
      Topn = 2,
      SampleIDs = ss,
      annotation_colors = coldata_palettes(htgs),
      arrangeby = c("Hrs", "sh"),
      annotations = c("Hrs", "sh"),
      cellheight = 10,
      cellwidth = 8
    )
  
  
  
  
  
  # label by alias in mapper
  
  htgs \%>\%
    DEphe(
      Topn = 3,
      SampleIDs = ss,
      annotation_colors = coldata_palettes(htgs),
      arrangeby = c("Hrs", "sh"),
      annotations = c("Hrs", "sh"),
      label_by = "ensembl_id",
      cellheight = 10,
      cellwidth = 8
    )
  
 
  
}
}
