% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpsPlot.R
\name{ExpsPlot_Table}
\alias{ExpsPlot_Table}
\title{Converts ExpsPlot data to matrix format}
\usage{
ExpsPlot_Table(
  ggplot_obj = NULL,
  xVar = NULL,
  facets = NULL,
  name_col = "Feature",
  value_col = "value"
)
}
\arguments{
\item{ggplot_obj}{ggplot object}

\item{xVar}{String indicating the x variable. If NULL (default),
the first column name reported by coldata_ function will be used.}

\item{facets}{A set of variables or expressions quoted by \code{\link[ggplot2:vars]{vars()}}
and defining faceting groups on the rows or columns dimension.
The variables can be named (the names are passed to \code{labeller}).

For compatibility with the classic interface, can also be a
formula or character vector. Use either a one sided formula, \code{~a + b},
or a character vector, \code{c("a", "b")}.}

\item{name_col}{string for naming column with feature names,
default is "Feature". Options include columns in mapper slot.}

\item{value_col}{string for naming column with expression
values, default is "value".}
}
\description{
Converts ExpsPlot data to matrix format
}
