% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hotgenes_logo.R
\name{Hotgenes_logo}
\alias{Hotgenes_logo}
\alias{Hotgenes_logo_path}
\title{Hex logo for Hotgenes}
\usage{
Hotgenes_logo_path()
}
\description{
Hex logo for Hotgenes

path to package logo
}
\examples{
library(Hotgenes)

logo_file_path <- Hotgenes_logo_path()

file.exists(logo_file_path)



}
