% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSVA.R
\name{HotgeneSets}
\alias{HotgeneSets}
\title{Support functions for GSVA}
\usage{
HotgeneSets(
  Hotgenes = NULL,
  ExpressionSlots = NULL,
  geneSets = NULL,
  method = c("ssgsea"),
  kcdf = c("Gaussian"),
  minSize = 2,
  maxSize = Inf,
  MapperCol = "Feature",
  voomaGroup = NULL,
  contrast_matrix = NULL,
  use_weights = TRUE,
  use_vooma = FALSE,
  ...
)
}
\arguments{
\item{Hotgenes}{Hotgenes object.}

\item{ExpressionSlots}{name of normalized data to use.
If NULL (default), the first name returned by ExpressionSlots_
will be used.
See HotgenesObj.}

\item{geneSets}{The gene sets.  Must be one of the classes supported by
\code{\linkS4class{GsvaGeneSets}}.  For a list of these classes, see its help page using
\code{help(GsvaGeneSets)}.}

\item{method}{string for method to use for GSVA package. Options
include: c("ssgsea","gsva", "zscore", "plage").}

\item{kcdf}{Character vector of length 1 denoting the kernel to use during
the non-parametric estimation of the empirical cumulative distribution
function (ECDF) of expression levels across samples. The value \code{kcdf="auto"}
will allow GSVA to automatically choose one of the possible values. The
value \code{kcdf="Gaussian"} is suitable when input expression values are
continuous, such as microarray fluorescent units in logarithmic scale,
RNA-seq log-CPMs, log-RPKMs, or log-TPMs. When input expression values are
integer counts, such as those derived from RNA-seq experiments, then this
argument should be set to \code{kcdf="Poisson"}. When we do not want to use a
kernel approach for the estimation of the ECDF, then we should set
\code{kcdf="none"}.}

\item{minSize}{Numeric vector of length 1.  Minimum size of the resulting gene
sets after gene identifier mapping. By default, the minimum size is 1.}

\item{maxSize}{Numeric vector of length 1.  Maximum size of the resulting gene
sets after gene identifier mapping. By default, the maximum size is \code{Inf}.}

\item{MapperCol}{String matching a column name returned by
Mapper_ function, default is "Feature".}

\item{voomaGroup}{string indicating the column name to be
passed to voomaByGroup(). Default is NULL. Ignored when
use_vooma = FALSE.}

\item{contrast_matrix}{an optional matrix with contrasts to be passed to
\code{\link[limma]{contrasts.fit}}.}

\item{use_weights}{logical, if TRUE (default) gene set sizes applied to 
ebayes trend. If FALSE, trend set to FALSE}

\item{use_vooma}{logical, if TRUE gsva matrix
passed to vooma before lmFit. If FALSE (default),
gsva matrix converted to EList object
and directly passed lmFit.}

\item{...}{Additional parameters for \code{\link[GSVA]{gsva}}.}
}
\description{
Support functions for GSVA
}
\examples{
if(interactive()){
  # load package
  library(Hotgenes)
  
  dds_Hotgenes_dir <- system.file("extdata",
                                  paste0("dds_Hotgenes", ".RDS"),
                                  package = "Hotgenes",
                                  mustWork = TRUE
  )
  
  # from DESeq2
  htgs <- readRDS(dds_Hotgenes_dir)
  
  # msigdbr_wrapper ---------------------------------------------------------
  choice_set <- "CP:KEGG"
  choice_id <- "gene_symbol"
  
  gsList <- msigdbr_wrapper(
    species = "human",
    set = choice_set,
    gene_col = choice_id
  )
  
  
  names(gsList)
  
  # HotgeneSets -------------------------------------------------------------
  htgs \%>\% Mapper_()
  
  HotgeneSets_out <- HotgeneSets(
    Hotgenes = htgs,
    geneSets = gsList,
    kcdf = "Gaussian",
    method = "ssgsea",
    minSize = 2,
    maxSize = Inf
  )
  
  
  HotgeneSets_out 
  
}
}
