% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsHotgenes.R
\name{na_check}
\alias{na_check}
\title{checks a vector for empty strings, NA, and "NA"}
\usage{
na_check(string = NULL)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}
}
\description{
checks a vector for empty strings, NA, and "NA"
}
\details{
The string is first passed to the
\code{\link[stringr]{str_replace_na}} function to
convert NA into "".

Then \code{\link[stringr]{str_detect}} is used to
detect spaces and "NA" with: "^[[:space:]]*$|^NA$"
}
\examples{
this_string <- c("a", NA, "", "b", "NA", "NANA", "na")

na_check(this_string)
}
