% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_Alias_support.R
\name{parse_mapper_aliases}
\alias{parse_mapper_aliases}
\title{Returns a named list showing where the aliases are found}
\usage{
parse_mapper_aliases(Hotgenes = NULL, features = NULL)
}
\arguments{
\item{Hotgenes}{Hotgenes object.}

\item{features}{strings to check}
}
\description{
Returns a named list showing where the aliases are found
}
