% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsHotgenes.R
\name{required_DE_numeric_cols}
\alias{required_DE_numeric_cols}
\title{internal required DE table column names}
\usage{
required_DE_numeric_cols()
}
\description{
internal required DE table column names
}
