% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_support_utilities.R
\name{TabCondition}
\alias{TabCondition}
\title{Provides the condition for when a specific tab is active}
\usage{
TabCondition(tabsetPanel_id = "tabs", id = NULL)
}
\arguments{
\item{tabsetPanel_id}{string for tabsetPanel id prefix to use.
Default is "tabs". This is what controls the visibility of sidepanels.}

\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
string condition
}
\description{
Provides the condition for when a specific tab is active
}
\details{
When deploying an app with multiple Hotgenes UIs,
don't
}
\examples{
# Example TabCondition ----------------------------------------------------

TabCondition(
  tabsetPanel_id = "dataset",
  id = "Exps"
)


TabCondition(
  tabsetPanel_id = "tabs",
  id = "Exps"
)
}
