% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_support_utilities.R
\name{UI_download_file}
\alias{UI_download_file}
\alias{Server_download_file}
\title{shiny download handler as a module}
\usage{
UI_download_file(id = NULL, label = "download", ...)

Server_download_file(
  id = NULL,
  filename = NULL,
  content = NULL,
  input,
  output,
  session,
  ...
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{label}{The label that should appear on the button.}

\item{...}{additional arguments for \code{\link[shiny:downloadHandler]{shiny::downloadHandler()}}}

\item{filename}{A string of the filename, including extension, that the
user's web browser should default to when downloading the file; or a
function that returns such a string. (Reactive values and functions may be
used from this function.)}

\item{content}{A function that takes a single argument \code{file} that is a
file path (string) of a nonexistent temp file, and writes the content to
that file path. (Reactive values and functions may be used from this
function.)}

\item{input}{shiny server input parameter}

\item{output}{shiny server output parameter}

\item{session}{Session from which to make a child scope (the default should
almost always be used).}
}
\description{
shiny download handler as a module

shiny plot download handler
}
