% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsHotgenes.R
\name{update_object}
\alias{update_object}
\title{update_object updates object class}
\usage{
update_object(Hotgenes = NULL)
}
\arguments{
\item{Hotgenes}{Hotgenes object.}
}
\description{
update_object updates object class
}
