% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hotgenes_logo.R
\name{Hotgenes_logo}
\alias{Hotgenes_logo}
\alias{Hotgenes_logo_path}
\alias{embed_Hotgenes_logo}
\title{Hex logo for Hotgenes}
\usage{
Hotgenes_logo_path()

embed_Hotgenes_logo(
  logo_dir = Hotgenes_logo_path(),
  align = "center",
  height = "100px",
  style = "text-align: center;margin-bottom:10px;"
)
}
\arguments{
\item{logo_dir}{path to www dir in shiny app working directory.
Default is \code{\link[=Hotgenes_logo_path]{Hotgenes_logo_path()}}.}

\item{align}{string html tag}

\item{height}{string html tag}

\item{style}{string html tag}
}
\description{
Hex logo for Hotgenes

path to package logo

Adds logo file to shiny app directory
}
\examples{
library(Hotgenes)

logo_file_path <- Hotgenes_logo_path()

file.exists(logo_file_path)



}
